requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
deepgram-sdk==3.2.7
python-dotenv==1.0.0
groq==0.11.0
httpx==0.26.0
aiortc==1.9.0
av==12.3.0
numpy==1.26.4

